Get DDR data

The DTCC Data Repository is a registered U.S. swap data repository that allows market participants to fulfil their public disclosure obligations under U.S. legislation. This function will give you the ability to download trade-level data that is reported by market participants. The field names are (and is assumed to be) the same for each asset class.

ddr(date, asset_class, field_specs = ddr_field_specs())

ddr_field_specs()

Arguments

date
the date for which data is required as Date or DateTime object. Only the year, month and day elements of the object are used and it must of be length one.

asset_class
the asset class for which you would like to download trade data. Valid inputs are "CR" (credit), "IR" (rates), "EQ" (equities), "FX" (foreign exchange), "CO" (commodities). This must be a string.

field_specs
a valid column specification that is passed to readr::read_csv() with a default value provided by ddr_field_specs()

Value

a tibble that contains the requested data. If no data exists on that date, an empty tibble is returned.

References

DDR Real Time Dissemination Platform

Examples

# NOT RUN {
library("lubridate")
ddr(ymd(20170525), "IR") # Not empty
# }