Get DDR data

The DTCC Data Repository is a registered U.S. swap data repository that allows market participants to fulfil their public disclosure obligations under U.S. legislation. This function will give you the ability to download trade-level data that is reported by market participants. Column specs are inferred from all records in the file (i.e. guess_max is set to Inf when calling readr::read_csv).

Usage

ddr(date, asset_class, show_col_types = FALSE)

Arguments

date

the date for which data is required as Date or DateTime object. Only the year, month and day elements of the object are used and it must of be length one.

asset_class

the asset class for which you would like to download trade data. Valid inputs are "CR" (credit), "IR" (rates), "EQ" (equities), "FX" (foreign exchange), "CO" (commodities). This must be a string.

show_col_types

if FALSE (default), do not show the guessed column types. If TRUE always show the column types, even if they are supplied. If NULL only show the column types if they are not explicitly supplied by the col_types argument.

Value

a tibble that contains the requested data. If no data exists on that date, an empty tibble is returned.

References

DDR Real Time Dissemination Platform

Examples

if (FALSE) {
ddr(as.Date("2017-05-25"), "IR") # Not empty
ddr(as.Date("2020-12-01"), "CR") # Not empty
}